Move CALL statement test to its own test
