+++ Snowflake.Data.Tests/SFStatementTest.cs
@@ -14,7 +14,7 @@ namespace Snowflake.Data.Tests
      * Mock rest request test
      */
     [TestFixture]
-    class SFStatementTest : SFBaseTest
+    class SFStatementTest
     {
         [Test]
         [Ignore("StatementTest")]
@@ -70,65 +70,6 @@ namespace Snowflake.Data.Tests
                 SFBaseResultSet resultSet = statement.Execute(0, "SELECT 1", null, false);
                 expectServiceName += "a";
                 Assert.AreEqual(expectServiceName, sfSession.ParameterMap[SFSessionParameter.SERVICE_NAME]);
-            }
-        }
-
-        [Test]
-        public void TestCallStatement()
-        {
-            using (DbConnection conn = new SnowflakeDbConnection())
-            {
-                conn.ConnectionString = ConnectionString;
-                conn.Open();
-
-                using (DbCommand command = conn.CreateCommand())
-                {
-                    try
-                    {
-                        command.CommandText = "ALTER SESSION SET USE_STATEMENT_TYPE_CALL_FOR_STORED_PROC_CALLS=true";
-                        command.ExecuteNonQuery();
-                    }
-                    catch (SnowflakeDbException ex)
-                    {
-                        // Expected error - SqlState: 22023, VendorCode: 1006
-                        Assert.AreEqual("22023", ex.SqlState);
-                    }
-
-                    command.CommandText = "create or replace procedure\n"
-                        + "TEST_SP_CALL_STMT_ENABLED(in1 float, in2 variant)\n"
-                        + "returns string language javascript as $$\n"
-                        + "let res = snowflake.execute({sqlText: 'select ? c1, ? c2', binds:[IN1, JSON.stringify(IN2)]});\n"
-                        + "res.next();\n"
-                        + "return res.getColumnValueAsString(1) + ' ' + res.getColumnValueAsString(2) + ' ' + IN2;\n"
-                        + "$$;";
-                    command.ExecuteNonQuery();
-
-                    command.CommandText = "call TEST_SP_CALL_STMT_ENABLED(?, to_variant(?))";
-
-                    var p1 = command.CreateParameter();
-                    p1.ParameterName = "1";
-                    p1.DbType = DbType.Double;
-                    p1.Value = 1;
-                    command.Parameters.Add(p1);
-
-                    var p2 = command.CreateParameter();
-                    p2.ParameterName = "2";
-                    p2.DbType = DbType.String;
-                    p2.Value = "[2,3]";
-                    command.Parameters.Add(p2);
-
-                    DbDataReader reader = command.ExecuteReader();
-
-                    string result = "1 \"[2,3]\" [2,3]";
-                    while (reader.Read())
-                    {
-                        Assert.AreEqual(result, reader.GetString(0));
-                    }
-
-                    command.CommandText = "drop procedure if exists TEST_SP_CALL_STMT_ENABLED(float, variant)";
-                    command.ExecuteNonQuery();
-                }
-                conn.Close();
             }
         }
     }
